### Before
```
[info] Benchmark                          Mode  Cnt        Score       Error  Units
[info] FunctionsBenchmark.benchAggregate  avgt   20  2646845.608 ± 15481.001  ns/op
[info] FunctionsBenchmark.benchCombine    avgt   20  2244241.371 ±  8921.944  ns/op
[info] FunctionsBenchmark.benchMonoid     avgt   20  7115897.251 ± 51600.308  ns/op
[info] FunctionsBenchmark.benchReduce     avgt   20  2628645.814 ± 46589.656  ns/op
[info] FunctionsBenchmark.benchSemigroup  avgt   20  7237729.044 ± 31415.158  ns/op
```

### After
```
[info] Benchmark                          Mode  Cnt        Score       Error  Units
[info] FunctionsBenchmark.benchAggregate  avgt   20  2486561.601 ± 15809.799  ns/op
[info] FunctionsBenchmark.benchCombine    avgt   20  2164058.998 ± 15960.153  ns/op
[info] FunctionsBenchmark.benchMonoid     avgt   20  2419176.004 ± 35601.634  ns/op
[info] FunctionsBenchmark.benchReduce     avgt   20  2324575.817 ± 76884.608  ns/op
[info] FunctionsBenchmark.benchSemigroup  avgt   20  3211684.503 ± 19937.007  ns/op
```